+++ CHANGELOG.md
@@ -4,3 +4,6 @@
 * feat: Create sidebar navigation (2025-05-17)
 * feat: Creating DnD wiki project (2025-05-17)
 
+### Fix
+* fix: Sidebar out of bounds and false title (2025-05-18)
+